Custom tests added

+++ models/staging/erp/stg_erp__sales_order_detail.sql
@@ -13,8 +13,8 @@ with
             , cast(orderqty as int) as order_qty
             , cast(productid as int) as product_fk
             , cast(specialofferid as int) as spercial_offer_fk
-            , cast(unitprice as numeric) as unit_price
-            , cast(unitpricediscount as numeric) as unit_price_discount
+            , cast(unitprice as float) as unit_price
+            , cast(unitpricediscount as float) as unit_price_discount
             --, cast(rowguid as string) as row_guid
             --, cast(modifieddate as date) as modified_date
         from sales_order_detail
